Issue-9550 - Port ArraySort to function-arrays subcrate
